Split loading_list tests.rs into submodules
The tests.rs file was 1094 lines, exceeding the 1000-line project
limit. Split into tests/mod.rs, tests/item.rs, tests/state.rs,
tests/component.rs, and tests/events.rs following the same pattern
used by tree and multi_progress component test splits.
