+++ src/component/loading_list/tests/component.rs
@@ -1,218 +1,5 @@
 use super::*;
 
-#[derive(Clone, Debug, PartialEq)]
-struct TestItem {
-    id: u32,
-    name: String,
-}
-
-fn make_items() -> Vec<TestItem> {
-    vec![
-        TestItem {
-            id: 1,
-            name: "Item One".to_string(),
-        },
-        TestItem {
-            id: 2,
-            name: "Item Two".to_string(),
-        },
-        TestItem {
-            id: 3,
-            name: "Item Three".to_string(),
-        },
-    ]
-}
-
-// ========================================
-// ItemState Tests
-// ========================================
-
-#[test]
-fn test_item_state_error_message() {
-    let state = ItemState::Error("Test error".to_string());
-    assert_eq!(state.error_message(), Some("Test error"));
-
-    let ready = ItemState::Ready;
-    assert!(ready.error_message().is_none());
-}
-
-#[test]
-fn test_item_state_symbols() {
-    assert_eq!(ItemState::Ready.symbol(0), " ");
-    assert_eq!(ItemState::Error("".to_string()).symbol(0), "✗");
-    // Loading has animated symbols
-    assert!(!ItemState::Loading.symbol(0).is_empty());
-}
-
-#[test]
-fn test_item_state_styles() {
-    let theme = Theme::default();
-    assert_eq!(ItemState::Ready.style(&theme), theme.normal_style());
-    assert_eq!(ItemState::Loading.style(&theme), theme.warning_style());
-    assert_eq!(
-        ItemState::Error("".to_string()).style(&theme),
-        theme.error_style()
-    );
-}
-
-// ========================================
-// LoadingListItem Tests
-// ========================================
-
-#[test]
-fn test_list_item_new() {
-    let item = LoadingListItem::new("data", "Label");
-    assert_eq!(item.data(), &"data");
-    assert_eq!(item.label(), "Label");
-    assert!(item.is_ready());
-}
-
-#[test]
-fn test_list_item_set_label() {
-    let mut item = LoadingListItem::new("data", "Old");
-    item.set_label("New");
-    assert_eq!(item.label(), "New");
-}
-
-#[test]
-fn test_list_item_set_state() {
-    let mut item = LoadingListItem::new("data", "Label");
-    item.set_state(ItemState::Loading);
-    assert!(item.is_loading());
-
-    item.set_state(ItemState::Error("err".to_string()));
-    assert!(item.is_error());
-}
-
-#[test]
-fn test_list_item_data_mut() {
-    let mut item = LoadingListItem::new("original", "Label");
-    *item.data_mut() = "modified";
-    assert_eq!(item.data(), &"modified");
-}
-
-// ========================================
-// LoadingListState Tests
-// ========================================
-
-#[test]
-fn test_state_new() {
-    let state: LoadingListState<String> = LoadingListState::new();
-    assert!(state.is_empty());
-    assert!(state.selected().is_none());
-    assert!(state.show_indicators());
-}
-
-#[test]
-fn test_state_with_items() {
-    let items = make_items();
-    let state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    assert_eq!(state.len(), 3);
-    assert_eq!(state.items()[0].label(), "Item One");
-    assert_eq!(state.items()[1].label(), "Item Two");
-    assert_eq!(state.items()[2].label(), "Item Three");
-}
-
-#[test]
-fn test_state_with_title() {
-    let state: LoadingListState<String> = LoadingListState::new().with_title("My List");
-    assert_eq!(state.title(), Some("My List"));
-}
-
-#[test]
-fn test_state_with_indicators() {
-    let state: LoadingListState<String> = LoadingListState::new().with_indicators(false);
-    assert!(!state.show_indicators());
-}
-
-#[test]
-fn test_state_set_loading() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    state.set_loading(0);
-    assert!(state.items()[0].is_loading());
-}
-
-#[test]
-fn test_state_set_ready() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    state.set_loading(0);
-    state.set_ready(0);
-    assert!(state.items()[0].is_ready());
-}
-
-#[test]
-fn test_state_set_error() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    state.set_error(0, "Failed to load");
-    assert!(state.items()[0].is_error());
-    assert_eq!(
-        state.items()[0].state().error_message(),
-        Some("Failed to load")
-    );
-}
-
-#[test]
-fn test_state_counts() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    state.set_loading(0);
-    state.set_loading(1);
-    state.set_error(2, "Error");
-
-    assert_eq!(state.loading_count(), 2);
-    assert_eq!(state.error_count(), 1);
-    assert!(state.has_loading());
-    assert!(state.has_errors());
-}
-
-#[test]
-fn test_state_selected() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    state.set_selected(Some(1));
-    assert_eq!(state.selected(), Some(1));
-    assert_eq!(state.selected_item().unwrap().label(), "Item Two");
-    assert_eq!(state.selected_data().unwrap().id, 2);
-}
-
-#[test]
-fn test_state_selected_clamped() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    state.set_selected(Some(100)); // Too high
-    assert_eq!(state.selected(), Some(2)); // Clamped to last
-}
-
-#[test]
-fn test_state_get() {
-    let items = make_items();
-    let state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    assert!(state.get(0).is_some());
-    assert!(state.get(100).is_none());
-}
-
-#[test]
-fn test_state_clear() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    state.set_selected(Some(0));
-
-    state.clear();
-    assert!(state.is_empty());
-    assert!(state.selected().is_none());
-}
-
 // ========================================
 // Component Tests
 // ========================================
@@ -612,79 +399,6 @@ fn test_last_empty_list() {
     assert!(output.is_none());
 }
 
-#[test]
-fn test_get_mut() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    if let Some(item) = state.get_mut(0) {
-        item.set_label("Modified");
-    }
-    assert_eq!(state.items()[0].label(), "Modified");
-
-    assert!(state.get_mut(100).is_none());
-}
-
-#[test]
-fn test_items_mut() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    state.items_mut().push(LoadingListItem::new(
-        TestItem {
-            id: 4,
-            name: "Item Four".to_string(),
-        },
-        "Item Four",
-    ));
-
-    assert_eq!(state.len(), 4);
-}
-
-#[test]
-fn test_set_title() {
-    let mut state: LoadingListState<String> = LoadingListState::new();
-    assert!(state.title().is_none());
-
-    state.set_title(Some("New Title".to_string()));
-    assert_eq!(state.title(), Some("New Title"));
-
-    state.set_title(None);
-    assert!(state.title().is_none());
-}
-
-#[test]
-fn test_set_show_indicators() {
-    let mut state: LoadingListState<String> = LoadingListState::new();
-    assert!(state.show_indicators());
-
-    state.set_show_indicators(false);
-    assert!(!state.show_indicators());
-}
-
-#[test]
-fn test_set_loading_invalid_index() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-
-    // This should not panic
-    state.set_loading(100);
-    state.set_ready(100);
-    state.set_error(100, "Error");
-}
-
-#[test]
-fn test_spinner_animation_frames() {
-    let state = ItemState::Loading;
-    // Test all 10 spinner frames (Braille dots matching SpinnerStyle::Dots)
-    let expected = ["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"];
-    for (i, &expected_frame) in expected.iter().enumerate() {
-        assert_eq!(state.symbol(i), expected_frame);
-    }
-    // Frame 10 should wrap to frame 0
-    assert_eq!(state.symbol(10), state.symbol(0));
-}
-
 // ========================================
 // State Transition Edge Case Tests
 // ========================================
@@ -874,221 +588,4 @@ fn test_large_loading_list_navigation() {
     // Up from first wraps to last
     LoadingList::update(&mut state, LoadingListMessage::Up);
     assert_eq!(state.selected(), Some(99));
-}
-
-// ========================================
-// handle_event Tests
-// ========================================
-
-use crate::input::{Event, KeyCode};
-
-#[test]
-fn test_handle_event_up() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    state.set_focused(true);
-
-    let msg = LoadingList::<TestItem>::handle_event(&state, &Event::key(KeyCode::Up));
-    assert_eq!(msg, Some(LoadingListMessage::Up));
-}
-
-#[test]
-fn test_handle_event_down() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    state.set_focused(true);
-
-    let msg = LoadingList::<TestItem>::handle_event(&state, &Event::key(KeyCode::Down));
-    assert_eq!(msg, Some(LoadingListMessage::Down));
-}
-
-#[test]
-fn test_handle_event_select() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    state.set_focused(true);
-
-    let msg = LoadingList::<TestItem>::handle_event(&state, &Event::key(KeyCode::Enter));
-    assert_eq!(msg, Some(LoadingListMessage::Select));
-}
-
-#[test]
-fn test_handle_event_vim_keys() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    state.set_focused(true);
-
-    // 'k' -> Up
-    let msg = LoadingList::<TestItem>::handle_event(&state, &Event::char('k'));
-    assert_eq!(msg, Some(LoadingListMessage::Up));
-
-    // 'j' -> Down
-    let msg = LoadingList::<TestItem>::handle_event(&state, &Event::char('j'));
-    assert_eq!(msg, Some(LoadingListMessage::Down));
-}
-
-#[test]
-fn test_handle_event_ignored_when_unfocused() {
-    let items = make_items();
-    let state = LoadingListState::with_items(items, |i| i.name.clone());
-    // Not focused by default
-    let msg = LoadingList::<TestItem>::handle_event(&state, &Event::key(KeyCode::Up));
-    assert_eq!(msg, None);
-}
-
-// ========================================
-// dispatch_event Tests
-// ========================================
-
-#[test]
-fn test_dispatch_event() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    state.set_focused(true);
-
-    // Down dispatches Down message, which selects the first item
-    let output = LoadingList::<TestItem>::dispatch_event(&mut state, &Event::key(KeyCode::Down));
-    assert!(matches!(
-        output,
-        Some(LoadingListOutput::SelectionChanged(0))
-    ));
-    assert_eq!(state.selected(), Some(0));
-}
-
-// ========================================
-// Instance Method Tests
-// ========================================
-
-#[test]
-fn test_instance_methods() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    state.set_focused(true);
-
-    // instance handle_event
-    let msg = state.handle_event(&Event::key(KeyCode::Down));
-    assert_eq!(msg, Some(LoadingListMessage::Down));
-
-    // instance update
-    let output = state.update(LoadingListMessage::Down);
-    assert!(matches!(
-        output,
-        Some(LoadingListOutput::SelectionChanged(0))
-    ));
-    assert_eq!(state.selected(), Some(0));
-
-    // instance dispatch_event
-    let output = state.dispatch_event(&Event::key(KeyCode::Down));
-    assert!(matches!(
-        output,
-        Some(LoadingListOutput::SelectionChanged(1))
-    ));
-    assert_eq!(state.selected(), Some(1));
-}
-
-// ========================================
-// Disabled State Tests
-// ========================================
-
-#[test]
-fn test_with_disabled() {
-    let items = make_items();
-    let state = LoadingListState::with_items(items, |i| i.name.clone()).with_disabled(true);
-    assert!(state.is_disabled());
-
-    let items = make_items();
-    let state = LoadingListState::with_items(items, |i| i.name.clone()).with_disabled(false);
-    assert!(!state.is_disabled());
-}
-
-#[test]
-fn test_set_disabled() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    assert!(!state.is_disabled());
-    state.set_disabled(true);
-    assert!(state.is_disabled());
-    state.set_disabled(false);
-    assert!(!state.is_disabled());
-}
-
-#[test]
-fn test_disabled_prevents_handle_event() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone());
-    state.set_focused(true);
-    state.set_disabled(true);
-    let msg = LoadingList::<TestItem>::handle_event(&state, &Event::key(KeyCode::Down));
-    assert_eq!(msg, None);
-}
-
-#[test]
-fn test_disabled_prevents_navigation() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone()).with_disabled(true);
-    state.set_focused(true);
-
-    let output = LoadingList::<TestItem>::update(&mut state, LoadingListMessage::Down);
-    assert_eq!(output, None);
-    assert_eq!(state.selected(), None);
-
-    let output = LoadingList::<TestItem>::update(&mut state, LoadingListMessage::Up);
-    assert_eq!(output, None);
-
-    let output = LoadingList::<TestItem>::update(&mut state, LoadingListMessage::First);
-    assert_eq!(output, None);
-
-    let output = LoadingList::<TestItem>::update(&mut state, LoadingListMessage::Last);
-    assert_eq!(output, None);
-
-    let output = LoadingList::<TestItem>::update(&mut state, LoadingListMessage::Select);
-    assert_eq!(output, None);
-}
-
-#[test]
-fn test_disabled_allows_programmatic_state_changes() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone()).with_disabled(true);
-
-    // SetLoading should still work when disabled
-    let output = LoadingList::<TestItem>::update(&mut state, LoadingListMessage::SetLoading(0));
-    assert!(output.is_some());
-    assert!(state.get(0).unwrap().is_loading());
-
-    // SetReady should still work when disabled
-    let output = LoadingList::<TestItem>::update(&mut state, LoadingListMessage::SetReady(0));
-    assert!(output.is_some());
-    assert!(state.get(0).unwrap().is_ready());
-
-    // Tick should still work when disabled
-    let output = LoadingList::<TestItem>::update(&mut state, LoadingListMessage::Tick);
-    assert!(output.is_none());
-}
-
-#[test]
-fn test_disabled_dispatch_event_returns_none() {
-    let items = make_items();
-    let mut state = LoadingListState::with_items(items, |i| i.name.clone()).with_disabled(true);
-    state.set_focused(true);
-
-    let output = state.dispatch_event(&Event::key(KeyCode::Down));
-    assert_eq!(output, None);
-}
-
-#[test]
-fn test_disabled_default_is_false() {
-    let state = LoadingListState::<TestItem>::new();
-    assert!(!state.is_disabled());
-}
-
-#[test]
-fn test_builder_chaining_with_disabled() {
-    let items = make_items();
-    let state = LoadingListState::with_items(items, |i| i.name.clone())
-        .with_title("Test")
-        .with_indicators(true)
-        .with_disabled(true);
-    assert!(state.is_disabled());
-    assert_eq!(state.title(), Some("Test"));
-    assert!(state.show_indicators());
 }
